New sketch 3A3A1 - Flight control stick
Here is the flight controller script. It has been tested on the complete assembly, with the mechanical and electronic components compliant to OH specifications. The template for creating the sketch was applied following issue #41. A Readme text file has also been added as an appendix to the script to provide users with a clear understanding of the integrated features.